# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/clients/{id}/groups": {
            "get": {
                "description": "Позволяет получить все группы, в которых состоит клиент",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clients"
                ],
                "summary": "GetClientGroups",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "client_id",
                        "name": "input",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/clienthandlers.GetGroupsRequestDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/clients/{id}/students": {
            "get": {
                "description": "Позволяет получить всех студентов клиента",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Clients"
                ],
                "summary": "GetClientStudents",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "client_id",
                        "name": "input",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/clienthandlers.GetClientStudentsHandlerResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/leads/createandgroup": {
            "post": {
                "description": "Транзакция на создания лида с записью в lead_groups",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Leads"
                ],
                "summary": "LeadsAndGroups",
                "parameters": [
                    {
                        "description": "Данные для создания лида и привязке к группе",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/leadhandlers.%D0%A1reateLeadInputDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.LeadResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/leads/groupedleads": {
            "post": {
                "description": "Позволяет получить сгруппированных лидов школы",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Leads"
                ],
                "summary": "GroupedLeads",
                "parameters": [
                    {
                        "description": "Фильтры: group_id и school_id(из токена авто)",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/leadhandlers.GroupedLeadInputDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/leadhandlers.GroupedLeadOutputDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/student-clients/createandgetBO": {
            "post": {
                "description": "Нужен, чтобы создать связь, получив при этом в ответ StudentClients бизнес-объект StudentClients",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "StudentClient"
                ],
                "summary": "CreateRelation",
                "parameters": [
                    {
                        "description": "Обычный studentClient на вход",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.StudentClientCreateUpdateDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/bodtos.BoDTO_StudentClientsReponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/students/{id}/clients": {
            "get": {
                "description": "Получить всех клиентов студента с метаданными (is_payer, relation)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Students"
                ],
                "summary": "GetStudentClients",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID студента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/studenthandlers.GetStudentClientsResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "bodtos.BoDTO_StudentClientsReponse": {
            "type": "object",
            "properties": {
                "birthdate": {
                    "type": "string"
                },
                "contact": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GroupResponseDTO"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "is_payer": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "relation": {
                    "type": "string"
                },
                "relation_id": {
                    "type": "integer"
                },
                "school_id": {
                    "type": "string"
                }
            }
        },
...
        "clienthandlers.GetClientStudentsHandlerResponseDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.StudentResponseDTO"
                    }
                }
            }
        },
...
        "clienthandlers.GetGroupsRequestDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GroupResponseDTO"
                    }
                }
            }
        },
...
        "dto.GroupResponseDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "entity_type": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "school_id": {
                    "type": "string"
                }
            }
        },
        "dto.LeadCreateUpdateDTO": {
            "type": "object",
            "properties": {
                "converted_to_client_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "qualification": {
                    "type": "string"
                },
                "school_id": {
                    "type": "string"
                },
                "source_id": {
                    "type": "integer"
                },
                "status_id": {
                    "type": "integer"
                },
                "trial_date": {
                    "type": "string"
                }
            }
        },
...
        "dto.LeadResponseDTO": {
            "type": "object",
            "properties": {
                "converted_to_client_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "qualification": {
                    "type": "string"
                },
                "school_id": {
                    "type": "string"
                },
                "source_id": {
                    "type": "integer"
                },
                "status_id": {
                    "type": "integer"
                },
                "trial_date": {
                    "type": "string"
                }
            }
        },
...
        "dto.StudentClientCreateUpdateDTO": {
            "type": "object",
            "properties": {
                "client_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_payer": {
                    "type": "boolean"
                },
                "relation": {
                    "type": "string"
                },
                "school_id": {
                    "type": "string"
                },
                "student_id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.StudentResponseDTO": {
            "type": "object",
            "properties": {
                "birthdate": {
                    "type": "string"
                },
                "contact": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "school_id": {
                    "type": "string"
                },
                "skill_level": {
                    "type": "string"
                }
            }
        },
...
        "leadhandlers.GroupedLeadInputDTO": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "integer"
                }
            }
        },
        "leadhandlers.GroupedLeadOutputDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.LeadCreateUpdateDTO"
                    }
                }
            }
        },
...
        "studenthandlers.GetStudentClientsResponseDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/bodtos.BoDTO_StudentClientsReponse"
                    }
                }
            }
        },
...
    }
}