# internal/Infrastructure/adapters/http/gin/bo_dtos/studentClientDTO.go
package bodtos

import "github.com/1DamnDaniel3/rscrm_go_serv/internal/Infrastructure/dto"

type BoDTO_StudentClientsReponse struct {
	Relation_id int64 `json:"relation_id"`
	ClientAndGroups
	Is_payer bool   `json:"is_payer"`
	Relation string `json:"relation"`
}

type ClientAndGroups struct {
	*dto.ClientResponseDTO
	Groups []dto.GroupResponseDTO `json:"groups"`
}